+++ demos/schemabrowser/src/main/java/is/codion/framework/demos/schemabrowser/client/ui/SchemaBrowserAppPanel.java
@@ -12,19 +12,26 @@ import is.codion.framework.demos.schemabrowser.domain.SchemaBrowser.Schema;
 import is.codion.framework.demos.schemabrowser.domain.SchemaBrowser.Table;
 import is.codion.framework.demos.schemabrowser.domain.SchemaBrowser.TableColumn;
 import is.codion.swing.common.ui.component.table.FilteredTable;
+import is.codion.swing.common.ui.laf.LookAndFeelProvider;
 import is.codion.swing.framework.model.SwingEntityApplicationModel;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityApplicationPanel;
 import is.codion.swing.framework.ui.EntityPanel;
 import is.codion.swing.framework.ui.EntityTablePanel;
 
+import com.formdev.flatlaf.intellijthemes.FlatAllIJThemes;
+
 import javax.swing.JTable;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
 import static is.codion.swing.framework.ui.TabbedDetailLayout.splitPaneResizeWeight;
+import static is.codion.swing.framework.ui.WindowDetailLayout.windowDetailLayout;
 
 public class SchemaBrowserAppPanel extends EntityApplicationPanel<SchemaBrowserAppPanel.SchemaBrowserApplicationModel> {
+
+  private static final String DEFAULT_FLAT_LOOK_AND_FEEL = "com.formdev.flatlaf.intellijthemes.FlatArcIJTheme";
 
   public SchemaBrowserAppPanel(SchemaBrowserApplicationModel applicationModel) {
     super(applicationModel);
@@ -40,7 +47,8 @@ public class SchemaBrowserAppPanel extends EntityApplicationPanel<SchemaBrowserA
 
     EntityPanel schemaPanel = new EntityPanel(schemaModel,
             config -> config.detailLayout(splitPaneResizeWeight(0.3)));
-    EntityPanel tablePanel = new EntityPanel(tableModel);
+    EntityPanel tablePanel = new EntityPanel(tableModel,
+            config -> config.detailLayout(windowDetailLayout()));
     EntityPanel columnPanel = new EntityPanel(columnModel);
     EntityPanel constraintPanel = new EntityPanel(constraintModel);
     EntityPanel columnConstraintPanel = new EntityPanel(columnConstraintModel);
@@ -56,12 +64,15 @@ public class SchemaBrowserAppPanel extends EntityApplicationPanel<SchemaBrowserA
   }
 
   public static void main(String[] args) {
+    Arrays.stream(FlatAllIJThemes.INFOS)
+            .forEach(LookAndFeelProvider::addLookAndFeelProvider);
     FilteredTable.AUTO_RESIZE_MODE.set(JTable.AUTO_RESIZE_ALL_COLUMNS);
     EntityTablePanel.CONDITION_PANEL_VISIBLE.set(true);
     EntityApplicationPanel.builder(SchemaBrowserApplicationModel.class, SchemaBrowserAppPanel.class)
             .applicationName("Schema Browser")
             .domainType(SchemaBrowser.DOMAIN)
             .defaultLoginUser(User.parse("scott:tiger"))
+            .defaultLookAndFeelClassName(DEFAULT_FLAT_LOOK_AND_FEEL)
             .start();
   }